add router support to server
