+++ server/src/index.ts
@@ -1,8 +1,8 @@
-import express, { Request, Response } from 'express';
+import express from 'express';
 import fs from 'fs';
 import cors from 'cors';
 import path from 'path';
-import { Campsite } from './types/campsite';
+import campsitesRouter from './routes/campsites';
 
 const app = express();
 const PORT = process.env.PORT || 3000;
@@ -10,6 +10,7 @@ const PORT = process.env.PORT || 3000;
 app.use(cors());
 app.use(express.json());
 
+// Serve static frontend if built
 const staticPath = path.join(__dirname, "../client/dist");
 const indexHtmlPath = path.join(staticPath, "index.html");
 
@@ -22,52 +23,8 @@ if (fs.existsSync(indexHtmlPath)) {
   console.warn("client/dist/index.html not found, skipping static route handling.");
 }
 
-app.get('/api/v1/campsites', (_req: Request, res: Response) => {
-  const filePath = path.join(__dirname, '../data/campsites.json');
-
-  fs.readFile(filePath, 'utf8', (err, data) => {
-    if (err) {
-      console.error('Error reading campsites.json:', err);
-      return res.status(500).json({ error: 'Unable to load campsites' });
-    }
-
-    try {
-      const campsites: Campsite[] = JSON.parse(data);
-      res.json(campsites);
-    } catch (parseError) {
-      console.error('Error parsing campsite data:', parseError);
-      res.status(500).json({ error: 'Invalid campsite data format' });
-    }
-  });
-});
-
-app.post('/api/v1/campsites', (req: Request, res: Response) => {
-  const filePath = path.join(__dirname, '../data/campsites.json');
-  const newCampsite: Campsite = req.body;
-
-  fs.readFile(filePath, 'utf8', (err, data) => {
-    if (err) {
-      console.error('Error reading campsites.json:', err);
-      return res.status(500).json({ error: 'Unable to load campsites' });
-    }
-
-    try {
-      const campsites: Campsite[] = JSON.parse(data);
-      campsites.push(newCampsite);
-
-      fs.writeFile(filePath, JSON.stringify(campsites, null, 2), (writeErr) => {
-        if (writeErr) {
-          console.error('Error writing to campsites.json:', writeErr);
-          return res.status(500).json({ error: 'Unable to save campsite' });
-        }
-        res.status(201).json(newCampsite);
-      });
-    } catch (parseError) {
-      console.error('Error parsing campsite data:', parseError);
-      res.status(500).json({ error: 'Invalid campsite data format' });
-    }
-  });
-});
+// API routes
+app.use('/api/v1/campsites', campsitesRouter);
 
 export default app;
